(#11417) Add EL 6 resource to configure pe-ruby-devel
The whole point of this work and this module is for me to easily get
pe-ruby-devel on the PE system I'm developing on in order to build the
jenkins.rb gem for use with Puppet Enterprise.

To reach this goal, this patch does a number of things.  First, it
manages basic "build essentials" on my CentOS 6.1 system.  gcc, make,
glibc-devel, curl, wget, etc...

Next, the module manages the epel-release package so that we have
additional 3rd party packages available to us.

In the same vein, the module manages
/etc/yum.repos.d/puppetenterprise.repo  This is where Mike might flip
his lid, but I can't wait for him to write this module.  I've configured
the module to use a base URL of
http://links.puppetlabs.com/puppet-enterprise which for the moment uses
a 302 redirect to faro.puppetlabs.lan.  This works great with curl -L
and yum.  In the future we can simply update the link to point to
something public if we choose to do so.

For the time being this module will be BROKEN for anyone who is not
inside the Puppet Labs local area network.  =(

Once the repository is configured, the module uses a normal package
resource to manage pe-ruby-devel which is the whole point of this
effort.

It works:

    notice: /Stage[main]/Pe_devel::Redhat/Package[pe-ruby-devel]/ensure:
    created

Spec tests are included with this patch to validate the desired behavior
described in the ticket:

    % rake spec
    (in /Users/jeff/vms/puppet/modules/pe_devel)

    pe_devel
      should create Class[pe_devel]
      on redhat el6 os families
        should create Class[pe_devel::redhat]
        should create Package[gcc]
        should create Package[wget]
        should create Package[curl]
        should create Package[glibc-devel]
        should create Package[pe-ruby-devel]
      on debian os families
        should create Notify[pe_devel unimplemented]
        should not create Class[pe_devel::redhat]

    Finished in 0.64855 seconds
    9 examples, 0 failures